Removed the commit sign-off section
This is only used in go-ipfs. I have moved it to the contribution guidelines for that repository, and will open a PR on that immediately. I ran into at least one contributor who was intimidated by this section, and it is not currently mandatory for all repositories (although it potentially should be).

Done as part of https://github.com/ipfs/community/issues/63